fix: test using Gradle build action to see if it speeds up builds

+++ examples/echo-kotlin/src/main/kotlin/ftl/echo/Echo.kt
@@ -7,13 +7,14 @@ import xyz.block.ftl.Context
 import xyz.block.ftl.Verb
 
 data class EchoRequest(val name: String)
+
 data class EchoResponse(val message: String)
 
 class Echo {
   @Verb
   fun echo(context: Context, req: EchoRequest): EchoResponse {
     val response = context.call(Time::time, TimeRequest())
-    val time = Instant.fromEpochSeconds(response.time.toLong(), 0)
+    val time = Instant.fromEpochSeconds(response.time, 0)
     return EchoResponse(message = "Hello, ${req.name}! The time is ${time}.")
   }
 }
